requirements.txt
pandas==1.3.2
scikit-learn==0.24.2
torch==1.9.0
torchvision==0.10.0
matplotlib==3.4.3
Flask==2.0.1
Pillow==7.0.0
requests==2.26.0